refactor: JPA 엔티티 매핑 정리

+++ src/main/java/com/yalmung/mini/commerce/product/model/Product.java
@@ -12,7 +12,7 @@ public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="product_no")
-    private Long productNo;
+    private Integer productNo;
 
     @Column(name="code")
     private String code;
